requirements.txt
apache-flink==2.0.0
apache-flink-libraries==2.0.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
jupyter==1.0.0
ipywidgets==8.1.1
transformers==4.30.2
torch==2.1.2 